failureRate/kubeflow_pipelines/preprocess/requirement.txt
Cython>=0.22
cmdstanpy==0.9.68
pystan==2.19.1.1
numpy>=1.15.4
pandas>=1.0.4
matplotlib>=2.0.0
plotly>=4.13
prophet
torch>=1.1.0
scikit-learn==1.0.1
salesforce-merlion==1.1.1
kubernetes